requirements.txt
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
pycparser==2.21
PyJWT==2.3.0
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3